Fixed ROS build warnings

+++ ros2/src/camera_ros/src/camera_subscriber_to_video.cpp
@@ -58,6 +58,6 @@ void Camera_subscriber_to_video::camera_image_callback(const sensor_msgs::msg::I
         return;
     }
 
-    cv::Mat camera_image = cv::Mat{ data->height, data->width, CV_8UC3, data->data.data() }.clone();
+    cv::Mat camera_image = cv::Mat{ (int)data->height, (int)data->width, CV_8UC3, data->data.data() }.clone();
     video_writer.write(camera_image);
 }
